20_vibecoding/Ydata/pyproject.toml
[project]
name = "ydata-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.0",
    "pandas>=2.3.3",
    "ydata-profiling>=4.17.0",
    "ydata-sdk[docx,text]>=3.0.8",
]
